requirements.txt
asgiref==3.8.1
Django==5.1.5
djangorestframework==3.14.0
et_xmlfile==2.0.0
gunicorn==23.0.0
openpyxl==3.1.5
packaging==24.2
psycopg2==2.9.10
psycopg2-binary==2.9.10
python-decouple==3.8
pytz==2025.1
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2024.2
